Moving exposed properties check into own class
