+++ json-tck/src/main/java/net/java/html/js/tests/JavaScriptBodyTest.java
@@ -134,22 +134,6 @@ public class JavaScriptBodyTest {
         assertEquals("number", doubleType, "Expecting number type: " + doubleType);
     }
 
-    private static void assertNoProp(Object obj, String name, Object arg) {
-        Object prop = Bodies.get(obj, name);
-        assertNull(prop, "Expecting no value for property " + name + ", but was " + Bodies.typeof(prop, false));
-
-        try {
-            Object res = Bodies.invoke(obj, name, arg);
-            if (name.equals(res)) {
-                return;
-            }
-            fail("Invoking " + name + " on " + obj + " returned " + res);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-
-    }
-
     enum Two {
         ONE, TWO;
     }
@@ -399,132 +383,6 @@ public class JavaScriptBodyTest {
         assertEquals(nullAndUnknown, 1, "Only one slot");
     }
 
-    @KOTest
-    public void exposedPropertiesOfAJavaObject() {
-        Sum s = new Sum();
-        Object[] props = Bodies.forIn(s);
-
-        List<Object> all = Models.asList(props);
-        assertEquals(0, all.size(), "No own properties: " + all);
-    }
-
-
-    @KOTest
-    public void exposedEqualsOfAJavaObject() {
-        Sum s = new Sum();
-        assertNoProp(s, "equals", s);
-    }
-
-    @KOTest
-    public void exposedHashCodeOfAJavaObject() {
-        Sum s = new Sum();
-
-        assertNoProp(s, "hashCode", null);
-    }
-
-    @KOTest
-    public void exposedWaitOfAJavaObject() {
-        Sum s = new Sum();
-
-        assertNoProp(s, "wait", null);
-    }
-
-    @KOTest
-    public void exposedGetClassOfAJavaObject() {
-        Sum s = new Sum();
-
-        assertNoProp(s, "getClass", null);
-    }
-
-    @KOTest
-    public void exposedNotifyOfAJavaObject() {
-        Sum s = new Sum();
-
-        assertNoProp(s, "notify", null);
-    }
-
-    @KOTest
-    public void exposedNotifyAllOfAJavaObject() {
-        Sum s = new Sum();
-
-        assertNoProp(s, "notifyAll", null);
-    }
-
-    @KOTest
-    public void exposedEqualsOfAJavaArray() {
-        Object s = new Object[5];
-        assertNoProp(s, "equals", s);
-    }
-
-    @KOTest
-    public void exposedHashCodeOfAJavaArray() {
-        Object s = new Object[5];
-
-        assertNoProp(s, "hashCode", null);
-    }
-
-    @KOTest
-    public void exposedWaitOfAJavaArray() {
-        Object s = new Object[5];
-
-        assertNoProp(s, "wait", null);
-    }
-
-    @KOTest
-    public void exposedGetClassOfAJavaArray() {
-        Object s = new Object[5];
-
-        assertNoProp(s, "getClass", null);
-    }
-
-    @KOTest
-    public void exposedNotifyOfAJavaArray() {
-        Object s = new Object[5];
-
-        assertNoProp(s, "notify", null);
-    }
-
-    @KOTest
-    public void exposedNotifyAllOfAJavaArray() {
-        Object s = new Object[5];
-
-        assertNoProp(s, "notifyAll", null);
-    }
-
-    @KOTest
-    public void exposedEqualsOfAJavaPrimitiveArray() {
-        Object s = new int[5];
-        assertNoProp(s, "equals", s);
-    }
-
-    @KOTest
-    public void exposedHashCodeOfAJavaPrimitiveArray() {
-        Object s = new int[5];
-
-        assertNoProp(s, "hashCode", null);
-    }
-
-    @KOTest
-    public void exposedWaitOfAJavaPrimitiveArray() {
-        Object s = new int[5];
-
-        assertNoProp(s, "wait", null);
-    }
-
-    @KOTest
-    public void exposedGetClassOfAJavaPrimitiveArray() {
-        Object s = new int[5];
-
-        assertNoProp(s, "getClass", null);
-    }
-
-    @KOTest
-    public void exposedNotifyOfAJavaPrimitiveArray() {
-        Object s = new int[5];
-
-        assertNoProp(s, "notify", null);
-    }
-
     @KOTest public void problematicString() {
         String orig = Bodies.problematicString();
         String js = Bodies.problematicCallback();
@@ -654,23 +512,24 @@ public class JavaScriptBodyTest {
         }
         throw new AssertionError("Expecting " + b + " but found " + a);
     }
-    private static void fail(String msg) {
+
+    static void fail(String msg) {
         throw new AssertionError(msg);
     }
 
-    private static void assertTrue(boolean c, String msg) {
+    static void assertTrue(boolean c, String msg) {
         if (!c) {
             throw new AssertionError(msg);
         }
     }
 
-    private static void assertFalse(boolean c, String msg) {
+    static void assertFalse(boolean c, String msg) {
         if (c) {
             throw new AssertionError(msg);
         }
     }
 
-    private static void assertNull(Object o, String msg) {
+    static void assertNull(Object o, String msg) {
         if (o != null) {
             throw new AssertionError(msg);
         }
